uh did this work?

+++ core/config/engine.h
@@ -60,7 +60,7 @@ private:
 	double _process_step = 0;
 
 	int ips = 60;
-	int _physics_delta = (1.0 / 60.0);
+	double _physics_delta = (1.0 / 60.0);
 	double physics_jitter_fix = 0.5;
 	double _fps = 1;
 	int _max_fps = 0;
@@ -110,7 +110,7 @@ public:
 	virtual int get_physics_ticks_per_second() const;
 	
 	virtual void set_physics_delta(int ticks_per_second);
-	virtual int get_physics_delta() const;
+	virtual double get_physics_delta() const;
 
 	virtual void set_max_physics_steps_per_frame(int p_max_physics_steps);
 	virtual int get_max_physics_steps_per_frame() const;
